requirements.txt
django
qrcode[pil]
pillow
